Move ProxyString to core.payload
This is specifically used for payload so move it between payload
specific utilities.
