+++ pyanaconda/core/payload.py
@@ -15,6 +15,11 @@
 # License and may only be used or replicated with the express permission of
 # Red Hat, Inc.
 #
+from urllib.parse import quote, unquote
+
+from pyanaconda.core.i18n import _
+from pyanaconda.core.regexes import URL_PARSE
+from pyanaconda.core.util import ensure_str
 
 
 def parse_nfs_url(nfs_url):
@@ -38,3 +43,94 @@ def parse_nfs_url(nfs_url):
             host = s[0]
 
     return options, host, path
+
+
+class ProxyStringError(Exception):
+    pass
+
+
+# TODO: Add tests
+class ProxyString(object):
+    """ Handle a proxy url."""
+    def __init__(self, url=None, protocol="http://", host=None, port="3128",
+                 username=None, password=None):
+        """ Initialize with either url
+        ([protocol://][username[:password]@]host[:port]) or pass host and
+        optionally:
+
+        protocol    http, https, ftp
+        host        hostname without protocol
+        port        port number (defaults to 3128)
+        username    username
+        password    password
+
+        The str() of the object is the full proxy url
+
+        ProxyString.url is the full url including username:password@
+        ProxyString.noauth_url is the url without username:password@
+        """
+        self.url = ensure_str(url, keep_none=True)
+        self.protocol = ensure_str(protocol, keep_none=True)
+        self.host = ensure_str(host, keep_none=True)
+        self.port = str(port)
+        self.username = ensure_str(username, keep_none=True)
+        self.password = ensure_str(password, keep_none=True)
+        self.proxy_auth = ""
+        self.noauth_url = None
+
+        if url:
+            self.parse_url()
+        elif not host:
+            raise ProxyStringError(_("No host url"))
+        else:
+            self.parse_components()
+
+    def parse_url(self):
+        """ Parse the proxy url into its component pieces
+        """
+        # NOTE: If this changes, update tests/regex/proxy.py
+        #
+        # proxy=[protocol://][username[:password]@]host[:port][path][?query][#fragment]
+        # groups (both named and numbered)
+        # 1 = protocol
+        # 2 = username
+        # 3 = password
+        # 4 = host
+        # 5 = port
+        # 6 = path
+        # 7 = query
+        # 8 = fragment
+        m = URL_PARSE.match(self.url)
+        if not m:
+            raise ProxyStringError(_("malformed URL, cannot parse it."))
+
+        # If no protocol was given default to http.
+        self.protocol = m.group("protocol") or "http://"
+
+        if m.group("username"):
+            self.username = ensure_str(unquote(m.group("username")))
+
+        if m.group("password"):
+            self.password = ensure_str(unquote(m.group("password")))
+
+        if m.group("host"):
+            self.host = m.group("host")
+            if m.group("port"):
+                self.port = m.group("port")
+        else:
+            raise ProxyStringError(_("URL has no host component"))
+
+        self.parse_components()
+
+    def parse_components(self):
+        """ Parse the components of a proxy url into url and noauth_url
+        """
+        if self.username or self.password:
+            self.proxy_auth = "%s:%s@" % (quote(self.username or ""),
+                                          quote(self.password or ""))
+
+        self.url = self.protocol + self.proxy_auth + self.host + ":" + self.port
+        self.noauth_url = self.protocol + self.host + ":" + self.port
+
+    def __str__(self):
+        return self.url
