+++ pyanaconda/payload/dnf/repomd.py
@@ -21,7 +21,7 @@ from requests import RequestException
 
 from pyanaconda.anaconda_loggers import get_packaging_logger
 from pyanaconda.core import util, constants
-from pyanaconda.core.util import ProxyString, ProxyStringError
+from pyanaconda.core.payload import ProxyString, ProxyStringError
 from pyanaconda.payload.dnf.utils import USER_AGENT
 
 log = get_packaging_logger()
