+++ pyanaconda/ui/gui/spokes/installation_source.py
@@ -33,7 +33,8 @@ from pyanaconda.flags import flags
 from pyanaconda.core.i18n import _, N_, CN_
 from pyanaconda.payload.image import find_optical_install_media, find_potential_hdiso_sources, \
     get_hdiso_source_info, get_hdiso_source_description
-from pyanaconda.core.util import ProxyString, ProxyStringError, cmp_obj_attrs, id_generator
+from pyanaconda.core.payload import ProxyString, ProxyStringError
+from pyanaconda.core.util import cmp_obj_attrs, id_generator
 from pyanaconda.ui.communication import hubQ
 from pyanaconda.ui.helpers import InputCheck, InputCheckHandler
 from pyanaconda.ui.gui import GUIObject
